Amélioration de historique pour prendre en compte le fait que un fichier peut etre repris lors d'une nouvelle conversation

+++ src/main/java/com/DeRivasRabary/insa/History/HistoryManager.java
@@ -1,5 +1,6 @@
 package com.DeRivasRabary.insa.History;
 
+import com.DeRivasRabary.insa.model.packet.File;
 import com.DeRivasRabary.insa.model.packet.Message;
 import com.opencsv.CSVWriter;
 
@@ -12,17 +13,18 @@ public class HistoryManager {
     private char sep = ',';
 
     public HistoryManager(String ipDest) throws IOException {
-        //nomdufichier @IPdestinataire.csv
-        //contient historique des conversations
-        this.filename = ipDest+"-data.csv";
-
+        this.filename = ipDest + "-data.csv";
+        // if fichier exist ne pas le créer
+        java.io.File f = new java.io.File(filename);
+        if (!f.isFile()) {
             CSVWriter writer = new CSVWriter(new FileWriter(filename));
-
-            //close the writer
             writer.close();
+        }
+
     }
 
     public void addMessage(Message message) throws IOException {
+        // il existe une méthode constructeur csvwriter plus complète pour rendre le programme plus robuste si besoin
         CSVWriter writer = new CSVWriter(new FileWriter(filename, true));
         String messageStored = message.getPseudoEmmeteur() + sep + message.getDate() + sep + message.getData();
         String [] record = messageStored.split(",");
